Rename action integrate_between_layers to compute_integral and action average_between_layers to compute_average to reflect their extended functionality
